refactor: extract AppState window-management extension (#574, #430)
Third slice of #430 (AppState decomposition). First non-adjacent slice
(prior 430a #571 and 430b #575 were single-range). Extracts 5 window-
management/termination delegator methods from two non-adjacent clusters
into a new Sources/BugNarrator/AppState+WindowManagement.swift as
`extension AppState { ... }`.

**Cluster A** (AppState.swift lines 864-870, 7 lines, 183 bytes, SHA-256
4e668ab312113bd9a10f88fac1e6555546e9cdebb1399d2c416030c8d7fe657a):
openTranscriptHistory, openRecordingControls.

**Cluster B** (AppState.swift lines 882-892, 11 lines, 345 bytes, SHA-256
90f8b0c069bf9d5331dbdd9cdc9a3e11fd4253e8a33952f3c21776b550b20f32):
openSettings, requestApplicationTermination, applicationShouldTerminate.

**openRecordingControlsAndStartSession()** (async, at former line 872-880,
sits between the clusters) has real behavior (guard on
`.recording` phase + `await startSession()`) and STAYS in AppState.swift.

Applied the spec-review-verify-SHA rule (feedback_spec_review_verify_sha.md)
+ endpoint verification + per-range byte-diff-against-origin (both
clusters silent) + reverse-order deletion (882-893 first, then 864-871
per the #556 non-adjacent protocol). Both cluster SHAs match origin
independently.

**Dead-import cleanup**: applicationShouldTerminate's return type
NSApplication.TerminateReply is AppKit-defined, so the new file imports
AppKit. That was the ONLY AppKit symbol in AppState.swift; after the
move, AppState.swift's `import AppKit` becomes dead and is removed in
this same commit (verified: grep for NSApplication/NSWindow/NSAlert/
NSResponder/NSApp/NSViewController/NSMenuItem/NSStatusItem in
AppState.swift → 0 remaining references).

The extension does NOT re-annotate @MainActor — inherited from
AppState's class-level annotation per SE-0316. 662 unit tests pass
unchanged; docs-sync validation passes.

Pre-build spec reviewed by codex in a fresh session (4 findings folded:
SHA gate pre-verified both clusters against origin, reverse-order
deletion confirmed correct, AppKit-not-Foundation required, dead-import
removal in same commit).
